Rename header to title analyzer
